requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.3
django-address==0.2.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-filters==0.2.1
django-materialize==1.0.1a2
django-shortuuidfield==0.1.3
django-tables2==2.0.3
pytz==2018.7
shortuuid==0.5.0
six==1.11.0